Dependencies for testing, generic PostgresSession now DatabaseSession without any Postgres mappings. Akka libraries are removed.
