+++ application.conf
@@ -7,9 +7,9 @@ server = "http4s"
 session-timeout = "30 minutes"
 
 db {
-    host = "localhost"
-    port = 5432
-    name = "messenger"
+    driver = "org.postgresql.Driver"
+    url = "jdbc:postgresql://localhost:5432"
+    database = "messenger"
     user = "msg"
     password = "msg_passwd"
     migrations = "liquibase/changelog.xml"
