+++ src/main/scala/app/Messenger.scala
@@ -1,11 +1,11 @@
 package app
 
-import app.impl.{AkkaHttpServer, Http4sServer, ServerImpl}
+import app.impl.Http4sServer
 import app.init.Init
-import cats.syntax.applicative._
-import cats.effect.ExitCase._
-import cats.effect.{ExitCode, IO, IOApp}
+import app.model.ServerConfig
+import cats.effect.{ExitCase, ExitCode, IO, IOApp}
 import com.typesafe.scalalogging.LazyLogging
+import pureconfig.ConfigSource
 
 import scala.concurrent.ExecutionContext
 
@@ -17,24 +17,19 @@ object Messenger extends IOApp with LazyLogging {
   logger.info(s"Application config: $config")
 
   /** Server startup */
-  private val server: ServerImpl = config.server match {
+  private val server: IO[ExitCode] = config.server match {
     case "http4s" =>
-      logger.info("Starting http4s server")
-      new Http4sServer
+      logger.info("Starting http4s server. Push 'Ctrl + C` to stop server...")
+      new Http4sServer().server
 
-    case _ => throw new IllegalArgumentException("No such server type. Only 'http4s' has implementation")
+    case _ => throw new IllegalArgumentException("No such server type. Only 'http4s' supported")
   }
 
-  println("Server is started. Enter 'Ctrl + C` to stop server...")
-
-  override def run(args: List[String]): IO[ExitCode] = server.server.guaranteeCase {
-    case Canceled =>
-      server.stop()
-      Init.stop()
-      logger.info("Application aborted, resources closed").pure[IO]
-    case _ =>
-      server.stop()
-      Init.stop()
-      logger.info("Normal application exit").pure[IO]
+  override def run(args: List[String]): IO[ExitCode] = server.guaranteeCase {
+    case ExitCase.Canceled =>
+      IO {
+        logger.info("Server is shutting down")
+        Init.stop()
+      }
   }
 }
